+++ R/breadcrumb.R
@@ -19,6 +19,14 @@
 #'   bc_breadcrumb_item("Components", href = "#"),
 #'   bc_breadcrumb_item("Breadcrumb", current = TRUE)
 #' )
+#'
+#' bc_breadcrumb(
+#'   bc_breadcrumb_item("Docs", href = "#"),
+#'   bc_breadcrumb_item("Theming", current = TRUE),
+#'   separator = "dot"
+#' )
+#'
+#' bc_breadcrumb_separator("dot")
 bc_breadcrumb <- function(..., label = "Breadcrumb", separator = "chevron-right") {
   separator <- arg_match(separator, c("chevron-right", "dot"))
   check_string(label, allow_empty = FALSE)