+++ R/item.R
@@ -41,6 +41,17 @@ bc_item_sizes <- c("default", "sm", "xs")
 #'   ),
 #'   href = "#"
 #' )
+#'
+#' bc_item_group(
+#'   bc_item(
+#'     htmltools::tags$section(htmltools::tags$h3("Playground")),
+#'     role = "listitem"
+#'   ),
+#'   bc_item(
+#'     htmltools::tags$section(htmltools::tags$h3("Models")),
+#'     role = "listitem"
+#'   )
+#' )
 bc_item <- function(...,
                     href = NULL,
                     variant = "default",